requirements.txt
strawberry-graphql[fastapi]
uvicorn
requests
feedparser
python-dotenv